+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### Exonum core
 
+- `RawTransaction` now has own implementation of `fmt::Debug` trait instead
+  of `#[derive(Debug)]`. The template of `RawTransaction`’s debug message is
+  `Transaction { version: #, service_id: #, message_type: #, length: #, hash: Hash(###) }`
+
 - `ExecutionError::with_description` method now takes `Into<String>`
   instead of `String` which allows to pass `&str` directly. (#592)